intersection(-no-root)?.js to core/dom/layout/
