+++ src/core/dom/layout/intersection-no-root.js
