+++ src/core/dom/layout/intersection.js
