+++ test/unit/core/dom/layout/test-intersection.js
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
-import {measureIntersectionNoRoot} from '../../../src/utils/intersection-no-root';
+import {
+  intersectionEntryToJson,
+  measureIntersection,
+} from '#core/dom/layout/intersection';
 
 describes.fakeWin('utils/intersection', {}, (env) => {
   function getInObConstructorStub() {
@@ -56,27 +59,27 @@ describes.fakeWin('utils/intersection', {}, (env) => {
   });
 
   it('should measure intersection for an element', async () => {
-    const intersection = measureIntersectionNoRoot(el);
+    const intersection = measureIntersection(el);
     fireIntersections([{x: 100, target: el}]);
     expect(await intersection).eql({x: 100, target: el});
   });
 
   it('should dedupe multiple measures', async () => {
-    const measure1 = measureIntersectionNoRoot(el);
-    const measure2 = measureIntersectionNoRoot(el);
+    const measure1 = measureIntersection(el);
+    const measure2 = measureIntersection(el);
     expect(measure1).equal(measure2);
   });
 
   it('should not dedupe multiple measures with entries in between', async () => {
-    const measure1 = measureIntersectionNoRoot(el);
+    const measure1 = measureIntersection(el);
     fireIntersections([{x: 100, target: el}]);
-    const measure2 = measureIntersectionNoRoot(el);
+    const measure2 = measureIntersection(el);
 
     expect(measure1).not.equal(measure2);
   });
 
   it('should only use the latest entry', async () => {
-    const intersection = measureIntersectionNoRoot(el);
+    const intersection = measureIntersection(el);
     const firstEntry = {x: 0, target: el};
     const secondEntry = {x: 100, target: el};
 
@@ -88,8 +91,8 @@ describes.fakeWin('utils/intersection', {}, (env) => {
     const el2 = env.win.document.createElement('p');
     env.win.document.body.appendChild(el2);
 
-    const intersection1 = measureIntersectionNoRoot(el);
-    const intersection2 = measureIntersectionNoRoot(el2);
+    const intersection1 = measureIntersection(el);
+    const intersection2 = measureIntersection(el2);
 
     const firstEntry = {x: 0, target: el};
     const secondEntry = {x: 2, target: el2};
@@ -113,8 +116,8 @@ describes.fakeWin('utils/intersection', {}, (env) => {
       },
     };
 
-    const intersection1 = measureIntersectionNoRoot(el1);
-    const intersection2 = measureIntersectionNoRoot(el2);
+    const intersection1 = measureIntersection(el1);
+    const intersection2 = measureIntersection(el2);
     const firstEntry = {target: el1};
     const secondEntry = {target: el2};
     fireIntersections([firstEntry]);
@@ -122,5 +125,44 @@ describes.fakeWin('utils/intersection', {}, (env) => {
 
     expect(await intersection1).equal(firstEntry);
     expect(await intersection2).equal(secondEntry);
+  });
+
+  describe('intersectionEntryToJson', () => {
+    const zeros = {
+      left: 0,
+      right: 0,
+      width: 0,
+      height: 0,
+      top: 0,
+      bottom: 0,
+      x: 0,
+      y: 0,
+    };
+
+    it('clones an IntersectionObserverEntry', () => {
+      const entry = {
+        time: 0,
+        intersectionRatio: 0,
+        rootBounds: zeros,
+        intersectionRect: zeros,
+        boundingClientRect: zeros,
+      };
+      const json = intersectionEntryToJson(entry);
+      expect(entry).eql(json);
+      expect(entry).not.equal(json);
+    });
+
+    it('clones with null rootBounds', () => {
+      const entry = {
+        time: 0,
+        intersectionRatio: 0,
+        rootBounds: null,
+        intersectionRect: zeros,
+        boundingClientRect: zeros,
+      };
+      const json = intersectionEntryToJson(entry);
+      expect(entry).eql(json);
+      expect(entry).not.equal(json);
+    });
   });
 });
